cloud_run/requirements.txt
Flask==2.3.3
debugpy # Required for debugging.
py-cord[speed]==2.5.0
py-cord==2.5.0
ezcord
pytube
googlesearch-python
google-api-python-client
google.generativeai
python-dotenv
flask_sqlalchemy
gunicorn